+++ weather@mockturtl/files/weather@mockturtl/3.8/applet.js
@@ -1,5 +1,5 @@
 const { weatherApplet } = require('./weather-applet');
 
 function main(metadata, orientation, panel_height, instance_id) {
-    return new weatherApplet.main(metadata, orientation, panel_height, instance_id);
+    return weatherApplet.main(metadata, orientation, panel_height, instance_id);
 }