+++ weather@mockturtl/webpack.config.js
@@ -12,20 +12,42 @@ module.exports = {
     entry: './src/3_8/applet.ts',
     devtool: "eval-source-map",
     target: 'node', // without webpack renames 'global'
-	optimization: {
+    optimization: {
         minimize: true,
-		usedExports: true,
+        usedExports: true,
     },
     module: {
         rules: [
             {
+                test: /\.ts$/,
                 use: 'ts-loader',
                 exclude: /node_modules/,
             },
+            {	// Some imported libraries are higher than
+                // es2017 (like Luxon) so we transpile them.
+                // Ensures compatibility to libjs52 package (Mint 19-19.3)
+                test: /\.js$/,
+                include: /node_modules/,
+                use: {
+                    loader: "babel-loader",
+                    options: {
+                        presets: [
+                            [
+                                "@babel/env",
+                                {
+                                    "targets": {
+                                        "firefox": "52"
+                                    }
+                                }
+                            ]
+                        ]
+                    }
+                }
+            }
         ],
     },
     resolve: {
-        extensions: ['.tsx', '.ts', '.js'],
+        extensions: ['.ts', '.js'],
     },
     output: {
         path: path.resolve(__dirname, `files/${appletName}/${cinnamonVersion}/`),